Update PigDiceGame-output1.txt
In line 26, the turn ended and as the score 14 is not held, the total score is still zero and it gets assigned to the turn score. the scenario where turn score is saved by using hold and then it hits a one is shown in the other output text file.