Update PigDiceGame-output2.txt
Here in line 25 we have saved the turn score by holding it and it is saved in the total score which is 9, but we hit a one in line 31. As we have saved the total score by holding it the turn score will remain the previous total score which is 9. therefore in line 41 we have turn score being 11 rather than zero. i felt like the point of holding is to save the progress so rather than making the turn score zero when we hit a one, i have made it get assigned to the total score. i have made an assumption for this scenario.